+++ src/components/ui/generalNote.js
@@ -1,21 +1,24 @@
 // @ts-nocheck
 import { CustomWindowEvents } from '../../settings';
-import { GeneralNoteEnums } from '../../enums';
+import { GeneralNoteCodes, GeneralNoteEnums } from '../../enums';
 import { styleErrors } from '../common/styles/errors';
-import { IdService, CustomEventService } from '../../services';
+import { IdService, CustomEventService, ClassIdService } from '../../services';
 import { errorIcon } from '../common/styles/statusIcons/status';
 
 class GeneralNote extends HTMLElement {
     constructor() {
       super();
       this.shadow = this.attachShadow({ mode: 'closed' });
-      this.id = this.getAttribute('id') || 'errorNote';
-      this.text = this.getAttribute('text') || '';
-      this.status = this.getAttribute('status') || '';
-      this.code = this.getAttribute('code') || '';
-      this.recipe = this.getAttribute('recipe') || '';
-      this.useBack = this.getAttribute('use-back') || '0';
+      this.id = this.getAttribute('id') || 'generalNote';
+      this.closeBtn = 'close';
+      this.size = '';
+      this.text = '';
+      this.status = '';
+      this.code = '';
+      this.recipe = '';
+      this.useBack = '0';
       this.container = this.id;
+      this.sizes = {w: 42, wUnits: '%', h: 100 };
     }
   
     connectedCallback() {
@@ -24,20 +27,31 @@ class GeneralNote extends HTMLElement {
     }
 
     initForm() {
-      this.$elBtn = IdService.idAndClick('close', this.shadow, this.toggleInfo.bind(this));
       this.$container = IdService.id(this.container, this.shadow);
-
-      this.$close = IdService.idAndClick('close', this.shadow, this.close.bind(this));
+      this.$close = IdService.idAndClick(this.closeBtn, this.shadow, this.close.bind(this));
 
       CustomEventService.event(CustomWindowEvents.generalNote.open, (e) => {
-        this.text = e.detail?.value || this.text;
+        this.setProps(e.detail?.value);
+
+        this.toggleStatusCl(this.status);
         this.toggleInfo();
+        this.setCustomSize(); 
+        this.setCloseVis();
         this.$container.showModal();
       });
-
       CustomEventService.event(CustomWindowEvents.generalNote.close, () => {
         this.$container?.close();
       });
+    }
+
+    setProps(rowProps = '') {
+      const props = JSON.parse(rowProps);
+      this.status = props.status;
+      this.code = props.code;
+      this.text = props.text;
+      this.size = props.size;
+      this.recipe = props.recipe;
+      this.useBack = props.useBack;
     }
 
     disconnectedCallback() {
@@ -46,18 +60,12 @@ class GeneralNote extends HTMLElement {
       }
     }
 
-    static get observedAttributes() { 
-      return ['text', 'status']; 
-    }
-
-    attributeChangedCallback(name, oldValue, newValue) {
-      if (name === 'text' && oldValue !== newValue) {
-        this.text = newValue;
-      }
-      if (name === 'status' && oldValue !== newValue) {
-        this.status = newValue;
-        this.toggleStatusCl(this.status);
-        this.toggleInfo();
+    setCustomSize() {
+      if (this.size !== '') {
+        const sizes = JSON.parse(this.size);
+        const { w, wUnits, h } = sizes;
+        this.$container.style.width = `${w}${wUnits}`;
+        this.$container.style.height = `${h}px`;
       }
     }
 
@@ -94,15 +102,23 @@ class GeneralNote extends HTMLElement {
       }
     }
 
+    setCloseVis() {
+      const el = ClassIdService.id(this.closeBtn, this.shadow);
+      el.style.opacity = '1';
+      if (this.code === GeneralNoteCodes.networkLost) {
+        el.style.opacity = '0';
+      }
+    }
+
     render() {
       this.shadow.innerHTML = `
           <style>
             ${errorIcon(22, 22)}
 
             #${this.container} {
-              width: 42%;
               background-color: #f2f2f2;
-              height: 120px;
+              width: ${this.sizes.w}${this.sizes.wUnits};
+              height: ${this.sizes.h}px;
             }
 
             .message {
@@ -135,7 +151,7 @@ class GeneralNote extends HTMLElement {
           </style>
           <dialog id="${this.container}">
             <div class="close">
-              <action-button id="close" label="Close" type="action" />
+              <action-button id="${this.closeBtn}" label="Close" type="action" />
             </div>
             <div id="status" class="error"></div>
           </dialog>