+++ src/pages/additional/index.js
@@ -1,5 +1,6 @@
 import './game';
 import './writer';
+import './gallery';
 import './map';
 import './tabs';
 import './account';